+++ src/tests/MasterTests.cpp
@@ -121,6 +121,11 @@ TEST_F(MasterTest, Init) {
   // Do nothing extra here, its all done in the SetUp()-function
 }
 
+TEST_F(MasterTest, WorkerStateStartsAtIdle) {
+  EXPECT_EQ(UnderTest->getCurrentState(), Status::WorkerState::Idle);
+  EXPECT_EQ(UnderTest->getCurrentStateMetric(), 0);
+}
+
 TEST_F(MasterTest, DestinationFilenameFromRelativePath) {
   REQUIRE_CALL(*StatusReporter,
                useAlternativeStatusTopic(StartCmd.ControlTopic));
@@ -145,6 +150,7 @@ TEST_F(MasterTest, StartWritingSuccess) {
       .TIMES(1);
   UnderTest->startWriting(StartCmd);
   EXPECT_EQ(UnderTest->getCurrentState(), Status::WorkerState::Writing);
+  EXPECT_EQ(UnderTest->getCurrentStateMetric(), 1);
 }
 
 TEST_F(MasterTest, StartWritingFailureWhenWriting) {
@@ -153,6 +159,7 @@ TEST_F(MasterTest, StartWritingFailureWhenWriting) {
       .TIMES(1);
   UnderTest->startWriting(StartCmd);
   ASSERT_EQ(UnderTest->getCurrentState(), Status::WorkerState::Writing);
+  EXPECT_EQ(UnderTest->getCurrentStateMetric(), 1);
   EXPECT_THROW(UnderTest->startWriting(StartCmd), std::runtime_error);
 }